fix test

+++ src/himena_relion/_utils.py
@@ -428,6 +428,7 @@ def open_with_lock(
     lock_dir = pipeline_path.parent / ".relion_lock"
     each_wait = 0.05
     num_trial = int(wait_sec / each_wait) + 1
+    existed = lock_dir.exists()
     try:
         for _ in range(num_trial):
             try:
@@ -448,11 +449,12 @@ def open_with_lock(
                 yield f
         finally:
             # remove the lock
-            with suppress(Exception):
-                lock_dir.rmdir()
+            if not existed:
+                with suppress(Exception):
+                    lock_dir.rmdir()
             pipeline_path.touch()
     finally:
-        if lock_dir.exists():
+        if not existed:
             with suppress(Exception):
                 lock_dir.rmdir()
